pyproject.toml
[project]
name = "app-web-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "plotly>=6.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "streamlit-scroll-to-top>=0.0.4",
]
